Make install.sh safe to run multiple times

+++ install.sh
@@ -9,6 +9,9 @@ if [[ ! -f ~/.ssh/id_rsa ]]; then
 fi
 cat ~/.ssh/id_rsa.pub >> ~/.ssh/authorized_keys 
 
+if [[ ! -d ./kolla-ansible ]]; then
+  git clone https://github.com/openstack/kolla-ansible
+fi
 virtualenv kolla-venv
 if  [[ ! -s kolla-venv/lib/python2.7/site-packages/ ]]; then
   ln -s /usr/lib64/python2.7/site-packages/selinux/ kolla-venv/lib/python2.7/site-packages/
@@ -16,21 +19,20 @@ fi
 source kolla-venv/bin/activate
 pip install -U pip
 pip install -U setuptools
-if [[ ! -d ./kolla-ansible ]]; then
-  git clone https://github.com/openstack/kolla-ansible
-fi
 pip install ./kolla-ansible
 pip install ansible
 
 sudo mkdir -p /etc/kolla
 sudo chown $USER: -R /etc/kolla/
 cp -r etc/kolla/* /etc/kolla/
-cp kolla-venv/share/kolla-ansible/etc_examples/kolla/passwords.yml /etc/kolla/
 mkdir -p /etc/kolla/config/ironic
 wget -O /etc/kolla/config/ironic/ironic-agent.initramfs https://tarballs.openstack.org/ironic-python-agent/tinyipa/files/tinyipa-master.gz
 wget -O /etc/kolla/config/ironic/ironic-agent.kernel https://tarballs.openstack.org/ironic-python-agent/tinyipa/files/tinyipa-master.vmlinuz
 
-kolla-genpwd
+if [[ ! -e /etc/kolla/passwords.yml ]]; then
+  cp kolla-venv/share/kolla-ansible/etc_examples/kolla/passwords.yml /etc/kolla/
+  kolla-genpwd
+fi
 kolla-ansible -i /etc/kolla/inventory/all-in-one bootstrap-servers
 
 if ! groups | grep docker >/dev/null; then
